requirements.txt
# 核心框架
pyautogen>=0.2.0
langchain>=0.1.0

# LLM 与 transformers
transformers>=4.38.0
torch>=2.0.0
accelerate>=0.27.0

# 嵌入与检索
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4          # 如果有 GPU，可换 faiss-gpu

# 可视化与日志
matplotlib>=3.8.0
seaborn>=0.13.0
wandb>=0.16.0

# RL 微调
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# NLP 工具
spacy>=3.7.0
# 在终端运行：python -m spacy download en_core_web_sm

# 其他工具
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0